Move LayoutCalculator
