+++ shame/src/frontend/rust_types/type_layout/layoutable/mod.rs
@@ -16,7 +16,7 @@ pub(crate) mod align_size;
 pub(crate) mod builder;
 pub(crate) mod ir_compat;
 
-pub use align_size::{FieldOffsets, MatrixMajor, array_size, array_stride, array_align};
+pub use align_size::{FieldOffsets, MatrixMajor, LayoutCalculator, array_size, array_stride, array_align};
 pub use builder::SizedOrArray;
 
 /// Types that have a defined memory layout.
